Thunder Card Refactoring
