+++ src/WaterWizard.Server/CardAbilities.cs
@@ -1,6 +1,7 @@
 using System.Numerics;
 using LiteNetLib;
 using WaterWizard.Server.Card;
+using WaterWizard.Server.Card.environment;
 using WaterWizard.Server.Card.healing;
 using WaterWizard.Server.Card.utility;
 using WaterWizard.Server.handler;
@@ -11,9 +12,6 @@ namespace WaterWizard.Server;
 public static class CardAbilities
 {
     private static readonly Random random = new();
-    private static readonly float THUNDER_INTERVAL = 1.75f;
-    private static float thunderTimer = 0;
-
     public static void HandleAbility(
         CardVariant variant,
         GameState gameState,
@@ -92,7 +90,6 @@ public static class CardAbilities
                         $"[Server] Invalid target for {variant} at ({targetCoords.X}, {targetCoords.Y})"
                     );
                 }
-                return;
             }
         }
         else if (UtilityCardFactory.IsUtilityCard(variant))
@@ -130,6 +127,10 @@ public static class CardAbilities
                 }
             }
         }
+        else if (EnvironmentCardFactory.IsEnvironmentCard(variant))
+        {
+
+        }
 
         switch (variant)
         {
@@ -152,7 +153,7 @@ public static class CardAbilities
             case "permanent":
                 Console.WriteLine($"[Server] Activated Card: {variant}");
                 break;
-            default:
+            default: //Duration is a number
                 try
                 {
                     int duration = int.Parse(durationString);
@@ -258,7 +259,7 @@ public static class CardAbilities
         if (card.Type == CardType.Utility)
         {
             // Verwende die übergebenen Handler
-            utilityCardHandler.HandleUtilityCard(variant, targetCoords, caster, defender);
+            // utilityCardHandler.HandleUtilityCard(variant, targetCoords, caster, defender);
             return;
         }
 
@@ -310,24 +311,11 @@ public static class CardAbilities
     /// <param name="card">The Card whose effect gets activated</param>
     /// <param name="passedTime">The time since last activation of the card. 0 if first time. Direct relationship
     /// between degree of effect and passed time needs to be implemented.</param>
-    internal static void HandleActivationEffect(Cards card, float passedTime)
+    internal static void HandleActivationEffect(GameState gameState, Cards card, float passedTime)
     {
         if (card.Variant == CardVariant.Thunder)
         {
-            thunderTimer -= passedTime / 1000f; // Konvertiere zu Sekunden
-
-            if (thunderTimer <= 0)
-            {
-                // Erzeuge neue Donnereinschläge
-                Console.WriteLine(
-                    $"[Server] Thunder strikes! Time since last activation: {passedTime}ms"
-                );
-                thunderTimer = THUNDER_INTERVAL;
-
-                // TODO: Implementiere die Logik für den Donnereinschlag
-                // Hier müssen wir die Koordinaten an den Client senden
-                // und die Treffer überprüfen
-            }
+            ThunderCard.HandleActivationEffect(gameState, passedTime);
         }
         else
         {
